This commit was generated by cvs2svn to compensate for changes in r16623, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@16624 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/tap.h
@@ -79,10 +79,10 @@ typedef struct tap_s {
     struct tape_file_record_s *tap_file_record;
 
     /* Tape counter in machine-cycles/8 for even looong tapes */
-    long cycle_counter;
+    int cycle_counter;
 
     /* Tape length in machine-cycles/8 */
-    long cycle_counter_total;
+    int cycle_counter_total;
 
     /* Tape counter.  */
     int counter;
